refactor(redux/dialogs): convert to slices
