+++ assets/scripts/app/keyboard_commands.js
@@ -8,7 +8,7 @@ import { handleSegmentMoveCancel } from '../segments/drag_and_drop'
 import { handleSegmentResizeCancel } from '../segments/resizing'
 import { getSignInData, isSignedIn } from '../users/authentication'
 import { t } from '../locales/locale'
-import { showDialog } from '../store/actions/dialogs'
+import { showDialog } from '../store/slices/dialogs'
 import { addToast } from '../store/slices/toasts'
 import { handleUndo, handleRedo } from '../store/actions/undo'
 import store from '../store'
