+++ assets/scripts/store/reducers/menus.js
@@ -2,9 +2,9 @@ import {
   SHOW_MENU,
   CLEAR_MENUS,
   SHOW_GALLERY,
-  SHOW_DIALOG,
   START_PRINTING
 } from '../actions'
+import { showDialog } from '../slices/dialogs'
 
 const initialState = {
   activeMenu: null
@@ -18,7 +18,7 @@ const menus = (state = initialState, action) => {
       }
     case CLEAR_MENUS:
     case SHOW_GALLERY: // Whenever gallery is shown, hide menus.
-    case SHOW_DIALOG: // Whenever a dialog is shown, also hide menus.
+    case showDialog.type: // Whenever a dialog is shown, also hide menus.
     case START_PRINTING: // Whenever in print mode, also hide menus.
       return initialState
     default:
