+++ assets/scripts/store/slices/dialogs.test.js
@@ -1,0 +1,24 @@
+/* eslint-env jest */
+import dialogs, { showDialog, clearDialogs } from './dialogs'
+
+describe('dialogs reducer', () => {
+  const initialState = {
+    name: null
+  }
+
+  it('should handle initial state', () => {
+    expect(dialogs(undefined, {})).toEqual(initialState)
+  })
+
+  it('should handle showDialog()', () => {
+    expect(dialogs(initialState, showDialog('DIALOG_NAME'))).toEqual({
+      name: 'DIALOG_NAME'
+    })
+  })
+
+  it('should handle clearDialogs()', () => {
+    expect(dialogs(initialState, clearDialogs())).toEqual({
+      name: null
+    })
+  })
+})
