+++ assets/scripts/streets/StreetMetaAnalytics.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { FormattedMessage } from 'react-intl'
 import { getStreetCapacity, formatCapacity } from '../util/street_analytics'
-import { showDialog } from '../store/actions/dialogs'
+import { showDialog } from '../store/slices/dialogs'
 
 function StreetMetaAnalytics (props) {
   const street = useSelector((state) => state.street)
